Extract behaviour for Sengoku.AI
And make the current one random.
